+++ appData/src/gb/include/vm_sgb.h
@@ -5,7 +5,9 @@
 #include <gb/sgb.h>
 
 #include "vm.h"
-                                                                                          
+
+BANKREF_EXTERN(VM_SGB)
+
 void vm_sgb_transfer(DUMMY0_t dummy0, DUMMY1_t dummy1, SCRIPT_CTX * THIS) OLDCALL NONBANKED;
 
 #endif