+++ appData/src/gb/src/core/vm_music.c
@@ -5,6 +5,8 @@
 #include "vm_music.h"
 #include "music_manager.h"
 
+BANKREF(VM_MUSIC)
+
 void vm_music_play(SCRIPT_CTX * THIS, UBYTE track_bank, const TRACK_T * track, UBYTE loop) OLDCALL BANKED {
     THIS;
     music_global_mute_mask = 0;
@@ -26,7 +28,7 @@ void vm_music_mute(SCRIPT_CTX * THIS, UBYTE channels) OLDCALL BANKED {
 void vm_music_routine(SCRIPT_CTX * THIS, UBYTE routine, UBYTE bank, UBYTE * pc) OLDCALL BANKED {
     THIS;
     script_event_t * event = &music_events[routine & 0x03];
-    event->script_bank = bank; 
+    event->script_bank = bank;
     event->script_addr = pc;
 }