+++ src/app/music/MusicRoot.js
@@ -43,7 +43,7 @@ ipcRenderer.on("music-data", (event, d) => {
       });
       break;
     case "play":
-      player.play(d.song);
+      player.play(d.song, d.position);
       ipcRenderer.send("music-data-receive", {
         action: "log",
         message: "playing",