+++ src/components/music/helpers.ts
@@ -80,15 +80,11 @@ const patternCelltoString = (
   p: PatternCell,
   fields: PatternCellKey[] = ["note", "instrument", "effectcode", "effectparam"]
 ) => {
-  return `|${renderNote(
-    fields.includes("note") ? p.note : null
-  )}${renderInstrument(
-    fields.includes("instrument") ? p.instrument : null
-  )}...${renderEffect(
-    fields.includes("effectcode") ? p.effectcode : null
-  )}${renderEffectParam(
-    fields.includes("effectparam") ? p.effectparam : null
-  )}`;
+  return `|${fields.includes("note") ? renderNote(p.note) : "   "}${
+    fields.includes("instrument") ? renderInstrument(p.instrument) : "  "
+  }...${fields.includes("effectcode") ? renderEffect(p.effectcode) : " "}${
+    fields.includes("effectparam") ? renderEffectParam(p.effectparam) : "  "
+  }`;
 };
 
 export const parsePatternToClipboard = (
@@ -96,7 +92,10 @@ export const parsePatternToClipboard = (
   channelId?: number,
   selectedCells?: number[]
 ) => {
-  let parsed: string[] = [];
+  let parsed: string[] = [
+    "GBStudio hUGETracker Piano format compatible with...",
+    "ModPlug Tracker  XM",
+  ];
 
   if (!selectedCells) {
     parsed = pattern.map((p) => {
@@ -134,7 +133,10 @@ export const parsePatternFieldsToClipboard = (
   pattern: PatternCell[][],
   selectedFields: number[]
 ) => {
-  const parsed: string[] = [];
+  const parsed: string[] = [
+    "GBStudio hUGETracker paste format compatible with...",
+    "ModPlug Tracker  XM",
+  ];
 
   const w =
     (selectedFields[selectedFields.length - 1] - selectedFields[0]) % 16;
@@ -143,6 +145,8 @@ export const parsePatternFieldsToClipboard = (
   );
   const firstRow = Math.floor(selectedFields[0] / 16);
   const firstColumn = selectedFields[0] % 16;
+  //console.log("w " + w + " h " + h + " row " + firstRow +
+  //  " column " + firstColumn+ " selected " + selectedFields);
 
   for (let i = firstRow; i <= firstRow + h; i++) {
     let rowStr = "";
@@ -174,22 +178,26 @@ export const parseClipboardToPattern = (clipboard: string) => {
     .map((r, i) => {
       console.log(`ROW ${i}: `, r);
       if (r[0] === "|") {
-        const channel = r.substring(1).trim().split("|");
+        const channel = r.substring(1).split("|");
         return channel.map((c, j) => {
           console.log(`CELL ${j}:`, c);
           const patternCell = new PatternCell();
-          console.log(
-            ">>>",
+          const cellString = [
             c.substring(0, 3),
             c.substring(3, 5),
             c.substring(8, 9),
-            c.substring(9, 11)
-          );
-          const note = noteNames.indexOf(c.substring(0, 3));
-          patternCell.note = note === -1 ? null : note;
-          patternCell.instrument = strToInt(c.substring(3, 5), 10, -1);
-          patternCell.effectcode = strToInt(c.substring(8, 9), 16);
-          patternCell.effectparam = strToInt(c.substring(9, 11), 16);
+            c.substring(9, 11),
+          ];
+          // Send -9 to not change parameter when merging
+          const note = noteNames.indexOf(cellString[0]);
+          patternCell.note =
+            cellString[0] !== "   " ? (note === -1 ? null : note) : -9;
+          patternCell.instrument =
+            cellString[1] !== "  " ? strToInt(cellString[1], 10, -1) : -9;
+          patternCell.effectcode =
+            cellString[2] !== " " ? strToInt(cellString[2], 16) : -9;
+          patternCell.effectparam =
+            cellString[3] !== "  " ? strToInt(cellString[3], 16) : -9;
           return patternCell;
         });
       }